DM-10560: clean of image retrieval error messages

+++ src/firefly/java/edu/caltech/ipac/visualize/net/SloanDssImageGetter.java
@@ -4,16 +4,14 @@
 package edu.caltech.ipac.visualize.net;
 
 
+import edu.caltech.ipac.util.Assert;
+import edu.caltech.ipac.util.DataGroup;
+import edu.caltech.ipac.util.VoTableUtil;
 import edu.caltech.ipac.util.download.CacheHelper;
 import edu.caltech.ipac.util.download.FailedRequestException;
-import edu.caltech.ipac.util.download.FileRetrieveException;
 import edu.caltech.ipac.util.download.HostPort;
 import edu.caltech.ipac.util.download.NetworkManager;
 import edu.caltech.ipac.util.download.URLDownload;
-import edu.caltech.ipac.util.Assert;
-import edu.caltech.ipac.util.ClientLog;
-import edu.caltech.ipac.util.DataGroup;
-import edu.caltech.ipac.util.VoTableUtil;
 
 import java.io.File;
 import java.io.IOException;
@@ -71,23 +69,16 @@ public class SloanDssImageGetter {
                 URLDownload.getDataToFile(new URL(urlString), outFile, null, false, true);
             }
             else {
-                throw new FileRetrieveException("Area not covered",
-                                                "votable returned not results, probably area is not covered: ", "SDSS");
+                throw new FailedRequestException("SDSS: Area not covered",
+                                                "votable returned not results, probably area is not covered: ");
             }
 
 
         } catch (SocketTimeoutException timeOutE) {
-            if (outFile.exists() && outFile.canWrite()) {
-                outFile.delete();
-            }
-            throw new FailedRequestException(
-                    FailedRequestException.SERVICE_FAILED,
-                    "Timeout", timeOutE);
+            if (outFile.exists() && outFile.canWrite()) outFile.delete();
+            throw timeOutE;
         } catch (MalformedURLException me) {
-            ClientLog.warning(me.toString());
-            throw new FailedRequestException(
-                    FailedRequestException.SERVICE_FAILED,
-                    "Details in exception", me);
+            throw new FailedRequestException( "Invalid URL", "Details in exception", me );
         }
 
     }
